Adjust IntcodeParser to work with dynamic memory and larger numbers
- adjust IntcodeComputer class
- adjust TEST class
- adjust Amplifier class

+++ AdventOfCode/Computer/Instruction.cs
@@ -5,11 +5,11 @@ namespace AdventOfCode.Computer
     public class Instruction
     {
         public Opcode Opcode { get; set; }
-        public List<int> Parameters { get; set; }
+        public List<long> Parameters { get; set; }
 
         public Instruction()
         {
-            Parameters = new List<int>();
+            Parameters = new List<long>();
         }
 
         public int ParameterCount() => Parameters.Count + 1;
